simplify millisecond clock implementation and improve it with CLOCK_MONOTONIC
2009-07-25  Paolo Bonzini  <bonzini@gnu.org>

	* kernel/Time.st: Simplify millisecond clock since GNU Smalltalk
	has had LargeIntegers for, ehm, quite some time.

2009-07-25  Paolo Bonzini  <bonzini@gnu.org>

libgst:
2009-07-25  Paolo Bonzini  <bonzini@gnu.org>

	* libgst/sysdep.h: Return 64-bit millisecond clock.
	* libgst/sysdep.c: Return 64-bit millisecond clock, use monotonic
	clock_gettime if available.
	* libgst/prims.def: Return 64-bit millisecond clock.

+++ libgst/sysdep.h
@@ -131,7 +131,7 @@ extern void _gst_debug (void)
   ATTRIBUTE_HIDDEN;
 
 /* Returns the local time since midnight in milliseconds */
-extern unsigned _gst_get_milli_time (void)
+extern uint64_t _gst_get_milli_time (void)
   ATTRIBUTE_HIDDEN;
 
 /* Returns the time in seconds since midnight Jan 1, 1970 (standard UNIX
